Implemented the interaction between picked-up items and inventory system
