+++ Assets/Canvas.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f17d675bb88fe6a42b147ff00e3f005c
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
