+++ Assets/InventorySystem Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 830bf49c3bdc1db4ca286da077f83584
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
