+++ Assets/InventorySystem Scripts/itemSlot.cs.meta
