--- Assets/Item.cs
@@ -1,30 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Item : MonoBehaviour
-{
-    [SerializeField] private string itemName;
-    [SerializeField] private int quantity;
-    [SerializeField] private Sprite sprite;
-    [SerializeField] private Sprite background;
-
-    [SerializeField] private string itemDescription;
-
-    InventoryManager inventoryManager;
-    // Start is called before the first frame update
-    void Start()
-    {
-        inventoryManager = GameObject.Find("Canvas").GetComponent<InventoryManager>();
-    }
-
-    // Update is called once per frame
-    private void OnCollisionEnter(Collision collision)
-    {
-        if(collision.gameObject.tag == "Player")
-        {
-            inventoryManager.AddItem(itemName, quantity, sprite);
-            //Destroy(gameObject);
-        }
-    }
-}
